Combine React examples
