--- examples/disclosures/package.json
@@ -1,38 +1,0 @@
-{
-	"name": "tiptap-disclosure-example-react",
-	"private": true,
-	"version": "0.0.0",
-	"type": "module",
-	"packageManager": "pnpm@10.8.0",
-	"scripts": {
-		"preinstall": "npx only-allow pnpm",
-		"dev": "vite",
-		"build": "tsc -b && vite build",
-		"lint": "eslint .",
-		"preview": "vite preview"
-	},
-	"dependencies": {
-		"@namesakefyi/tiptap-extensions": "file:../../",
-		"@tiptap/extension-document": "^2.11.7",
-		"@tiptap/extension-placeholder": "^2.11.7",
-		"@tiptap/pm": "^2.11.7",
-		"@tiptap/react": "^2.11.7",
-		"@tiptap/starter-kit": "^2.11.7",
-		"react": "^19.0.0",
-		"react-dom": "^19.0.0"
-	},
-	"devDependencies": {
-		"@eslint/js": "^9.22.0",
-		"@namesakefyi/tiptap-extensions": "file:../../",
-		"@types/react": "^19.0.10",
-		"@types/react-dom": "^19.0.4",
-		"@vitejs/plugin-react": "^4.3.4",
-		"eslint": "^9.22.0",
-		"eslint-plugin-react-hooks": "^5.2.0",
-		"eslint-plugin-react-refresh": "^0.4.19",
-		"globals": "^16.0.0",
-		"typescript": "~5.7.2",
-		"typescript-eslint": "^8.26.1",
-		"vite": "^6.3.1"
-	}
-}
