--- examples/disclosures/src/index.css
@@ -1,72 +1,0 @@
-:root {
-  font-family: system-ui, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-  color: #111111;
-  background-color: #f1f1f1;
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-a {
-  font-weight: 500;
-  color: #646cff;
-  text-decoration: inherit;
-}
-
-a:hover {
-  color: #535bf2;
-}
-
-body {
-  margin: 0;
-  min-width: 320px;
-  min-height: 100vh;
-}
-
-h1 {
-  font-size: 2.5em;
-  line-height: 1.1;
-  margin-block-start: unset;
-}
-
-p {
-  margin: 0;
-}
-
-button {
-  border-radius: 0.8rem;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #dddddd;
-  color: #111111;
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-
-button:hover {
-  background-color: #cccccc;
-}
-
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-  outline-offset: 2px;
-}
-
-button:disabled {
-  pointer-events: none;
-  opacity: 0.5;
-}
-
-.tiptap {
-  margin-top: 1rem;
-  border: 1px solid #cccccc;
-  border-radius: 0.8rem;
-  padding: 1rem;
-}
