+++ examples/react/package.json
