+++ examples/react/tsconfig.app.json
