+++ examples/react/tsconfig.json
