+++ examples/react/vite.config.ts
